src/requirements.txt
langchain
python-dotenv
streamlit
google-generativeai
langchain-google-genai